create shared batchClient.
