+++ internal/component/common/loki/wal/writer.go
@@ -115,10 +115,8 @@ func NewWriter(walCfg Config, logger log.Logger, reg prometheus.Registerer) (*Wr
 }
 
 func (wrt *Writer) start(maxSegmentAge time.Duration) {
-	wrt.wg.Add(1)
 	// main WAL writer routine
-	go func() {
-		defer wrt.wg.Done()
+	wrt.wg.Go(func() {
 		for e := range wrt.entries {
 			if err := wrt.entryWriter.WriteEntry(e, wrt.wal, wrt.log); err != nil {
 				level.Error(wrt.log).Log("msg", "failed to write entry", "err", err)
@@ -135,17 +133,12 @@ func (wrt *Writer) start(maxSegmentAge time.Duration) {
 			}
 			wrt.writeSubscribersLock.RUnlock()
 		}
-	}()
+	})
 	// WAL cleanup routine that cleans old segments
-	wrt.wg.Add(1)
-	go func() {
-		defer wrt.wg.Done()
+	wrt.wg.Go(func() {
 		// By cleaning every 10th of the configured threshold for considering a segment old, we are allowing a maximum slip
 		// of 10%. If the configured time is 1 hour, that'd be 6 minutes.
-		triggerEvery := maxSegmentAge / 10
-		if triggerEvery < minimumCleanSegmentsEvery {
-			triggerEvery = minimumCleanSegmentsEvery
-		}
+		triggerEvery := max(maxSegmentAge/10, minimumCleanSegmentsEvery)
 		trigger := time.NewTicker(triggerEvery)
 		for {
 			select {
@@ -159,7 +152,7 @@ func (wrt *Writer) start(maxSegmentAge time.Duration) {
 				return
 			}
 		}
-	}()
+	})
 }
 
 func (wrt *Writer) Chan() chan<- loki.Entry {
